Final Revisions (Part 3)
Content Done! Responsiveness Done! Remaining Design Done!

Final bug: the cart icon drops into the dropdown on mobile when it should stay in the navbar.

Also, the indents of the code are super sloppy. I intend on changing those, but left them to the very end becuase the browser doesn't care about indents.